This commit contains the initial implementation that supports computation of relations (maps, constraints, etc.) in the stationary time level. A detailed documentation is also planned. In addition, the container class hierarchy is revised. Some bugs are fixed. We now have a separate "freeze" and "thaw" method for each type of container. The "remap" method merely remaps a container now. But there are still some implementation problems left in this commit.

+++ src/include/Loci_types.h
@@ -416,7 +416,7 @@ namespace Loci {
 
   template <class T,size_t n> inline std::ostream &
     operator<<(std::ostream &s, const Array<T,n> &v) {
-    for(int i=0;i<n;++i)
+    for(size_t i=0;i<n;++i)
       s << v[i] << ' ' ;
     return s ;
   }
